+++ NEWS.md
@@ -1,7 +1,8 @@
 # TBSignatureProfiler (development version)
 
 * Currently have 49 signatures available
-* Working on a function and vignette on how to submit a signature to our github repo
+* Added `addTBsignature()` to more easily facilitate updating signatures in package
+* 
 
 # TBSignatureProfiler 0.0.0.9000